// CovidApp/src/models/liveCountry.model.ts
export interface CountryLiveArray {
    CountriesLive: CountryLive[];
}

export interface CountryLive {
    Country: string;
    CountryCode: string;
    Province: string;
    City: string;
    CityCode: string;
    Lat: string;
    Lon: string;
    Confirmed: number;
    Deaths: number;
    Recovered: number;
    Active: number;
    Date: string;
}